+++ opentelemetry-sdk/src/metrics/internal/last_value.rs
@@ -4,10 +4,8 @@ use std::{
     time::SystemTime,
 };
 
-use crate::metrics::{
-    attributes::AttributeSet,
-    data::{self, Gauge},
-};
+use crate::attributes::AttributeSet;
+use crate::metrics::data::{self, Gauge};
 
 use super::{Aggregator, Number};