+++ opentelemetry-sdk/src/metrics/meter_provider.rs
@@ -1,12 +1,15 @@
 use core::fmt;
-use std::sync::{
-    atomic::{AtomicBool, Ordering},
-    Arc,
+use std::{
+    borrow::Cow,
+    sync::{
+        atomic::{AtomicBool, Ordering},
+        Arc,
+    },
 };
 
 use opentelemetry_api::{
     metrics::{noop::NoopMeterCore, InstrumentProvider, Meter as ApiMeter, MetricsError, Result},
-    Context,
+    Context, KeyValue,
 };
 
 use crate::{instrumentation::Scope, Resource};
@@ -73,13 +76,14 @@ impl MeterProvider {
 impl opentelemetry_api::metrics::MeterProvider for MeterProvider {
     fn versioned_meter(
         &self,
-        name: &'static str,
-        version: Option<&'static str>,
-        schema_url: Option<&'static str>,
+        name: Cow<'static, str>,
+        version: Option<Cow<'static, str>>,
+        schema_url: Option<Cow<'static, str>>,
+        attributes: Option<Vec<KeyValue>>,
     ) -> ApiMeter {
         let inst_provider: Arc<dyn InstrumentProvider + Send + Sync> =
             if !self.is_shutdown.load(Ordering::Relaxed) {
-                let scope = Scope::new(name, version, schema_url);
+                let scope = Scope::new(name, version, schema_url, attributes);
                 Arc::new(SdkMeter::new(scope, self.pipes.clone()))
             } else {
                 Arc::new(NoopMeterCore::new())