+++ opentelemetry-semantic-conventions/scripts/templates/header_resource.rs
@@ -9,12 +9,12 @@
 //!
 //! ```rust,no_run
 //! use opentelemetry::sdk;
-//! use opentelemetry_semantic_conventions as semcov;
+//! use opentelemetry_semantic_conventions as semconv;
 //!
-//! let _tracer = opentelemetry::sdk::export::trace::stdout::new_pipeline()
-//!     .with_trace_config(sdk::trace::config().with_resource(sdk::Resource::new(vec![
-//!         semcov::resource::SERVICE_NAME.string("my-service"),
-//!         semcov::resource::SERVICE_NAMESPACE.string("my-namespace"),
+//! let _tracer = sdk::trace::TracerProvider::builder()
+//!     .with_config(sdk::trace::config().with_resource(sdk::Resource::new(vec![
+//!         semconv::resource::SERVICE_NAME.string("my-service"),
+//!         semconv::resource::SERVICE_NAMESPACE.string("my-namespace"),
 //!     ])))
-//!     .install_simple();
+//!     .build();
 //! ```