+++ opentelemetry-zipkin/src/exporter/mod.rs
@@ -176,8 +176,9 @@ impl ZipkinPipelineBuilder {
         provider_builder = provider_builder.with_config(config);
         let provider = provider_builder.build();
         let tracer = provider.versioned_tracer(
-            "opentelemetry-zipkin",
-            Some(env!("CARGO_PKG_VERSION")),
+            "opentelemetry-zipkin".into(),
+            Some(env!("CARGO_PKG_VERSION").into()),
+            None,
             None,
         );
         let _ = global::set_tracer_provider(provider);
@@ -197,8 +198,9 @@ impl ZipkinPipelineBuilder {
         provider_builder = provider_builder.with_config(config);
         let provider = provider_builder.build();
         let tracer = provider.versioned_tracer(
-            "opentelemetry-zipkin",
-            Some(env!("CARGO_PKG_VERSION")),
+            "opentelemetry-zipkin".into(),
+            Some(env!("CARGO_PKG_VERSION").into()),
+            None,
             None,
         );
         let _ = global::set_tracer_provider(provider);